refactor: split msg and translator
