+++ packages/translate/test/msg.test.ts
@@ -1,6 +1,6 @@
 import { test } from 'node:test';
 import assert from 'node:assert/strict';
-import { msg, Translator } from '../src/index.ts';
+import { msg } from '../src/utils.ts';
 
 test('msg with string returns id and message', () => {
   assert.deepEqual(msg('hello'), { id: 'hello', message: 'hello' });
@@ -28,28 +28,6 @@ test('msg with template string returns placeholders and values', () => {
     message: 'Hello, ${0}!',
     values: [name],
   });
-});
-
-test('translator substitutes template values', () => {
-  const name = 'World';
-  const t = new Translator('en', {});
-  assert.equal(t.gettext(msg`Hello, ${name}!`), 'Hello, World!');
-});
-
-test('translator applies translations with placeholders', () => {
-  const name = 'World';
-  const translations = {
-    ru: {
-      translations: {
-        '': {
-          'Hello, ${0}!': { msgid: 'Hello, ${0}!', msgstr: ['Привет, ${0}!'] },
-        },
-      },
-    },
-  } as any;
-  const t = new Translator('en', translations);
-  t.useLocale('ru');
-  assert.equal(t.gettext(msg`Hello, ${name}!`), 'Привет, World!');
 });
 
 test('msg disallows computed strings', () => {
